+++ src/views/@common/hooks/useSetScreenSize.ts
@@ -1,0 +1,20 @@
+import { useEffect } from 'react';
+
+const useSetScreenSize = () => {
+  const setScreenSize = () => {
+    // window width 관련
+    const windowWidth = window.innerWidth || document.documentElement.clientWidth || document.body.clientWidth;
+    const maxWidth = Math.min(375, windowWidth);
+    document.documentElement.style.setProperty('--app-max-width', `${maxWidth}px`);
+  };
+
+  useEffect(() => {
+    setScreenSize();
+    window.addEventListener('resize', setScreenSize);
+    return () => {
+      window.removeEventListener('resize', setScreenSize);
+    };
+  }, []);
+};
+
+export default useSetScreenSize;
