--- src/views/@common/utils/removeToken.ts
@@ -1,5 +1,0 @@
-const removeToken = () => {
-  sessionStorage.clear();
-};
-
-export default removeToken;
